test(tests, pool_lp): add LP witness test + split unit and E2E tests in tests
