+++ tests/src/tests.rs
@@ -3,19 +3,11 @@ use crate::perun::random;
 use crate::perun::virtual_channel::*;
 
 use super::*;
-use alloy_sol_types::SolValue;
 use ckb_occupied_capacity::Capacity;
-use ckb_testtool::ckb_types::{bytes::Bytes, packed::*, prelude::*};
+use ckb_testtool::ckb_types::{packed::*, prelude::*};
 use ckb_testtool::context::Context;
 use perun;
 use perun::{test, virtual_channel};
-use perun_common::perun_types::{
-    Allocation, AnyBalances, AnyBalancesUnion, Balances, Bool, CKByteDistribution, ChannelState,
-    ETHAsset, ETHBalances, ETHDistribution, EthAddress, LockedBalances, SEC1EncodedPubKey,
-    SubAlloc,
-};
-use perun_common::sig::{ethereum_message_hash, verify_signature};
-use perun_common::sol::convert_ckb_state;
 use std::cell::RefCell;
 use std::rc::Rc;
 use std::sync::Mutex;
@@ -25,208 +17,6 @@ const CHALLENGE_DURATION_MS: u64 = 10 * 1000;
 
 // Include your tests here
 // See https://github.com/xxuejie/ckb-native-build-sample/blob/main/tests/src/tests.rs for more examples
-
-#[test]
-fn test_signature() {
-    // This tests the interoperability between the on-chain signature verification
-    // and the key generation & signing in the perun-ckb-backend's wallet.
-
-    // This signature was generated by the wallet in the perun-ckb-backend
-    let sig_string = "3044022041b97f9d915cb53805988c1125cb916bcb3c83dcd9fd4ea3cdfef7af4dc2c813022038a0e6cda8417467d5fe5502f69c7d3c8ec72db79d5c556b660082da65aa18c3";
-    let sig = hex::decode(sig_string).expect("decoding signature");
-    let sig_bytes: Bytes = sig.into();
-
-    // This public key was generated by the wallet in the perun-ckb-backend
-    let pubkey_string = "032d4269f527a6b58ea5fd43910ac515e09b9364fc053f72ffa7b4572a1f3be9a7";
-    let pubkey = hex::decode(pubkey_string).expect("decoding pubkey");
-    let pubkey_bytes: [Byte; 33] = pubkey
-        .iter()
-        .map(|x| (*x).into())
-        .collect::<Vec<Byte>>()
-        .try_into()
-        .unwrap();
-    SEC1EncodedPubKey::new_builder().set(pubkey_bytes).build();
-
-    let balances_array: [Uint64; 2] = [10u64.pack(), 11u64.pack()];
-    let balances = Balances::new_builder()
-        .assets(
-            Allocation::new_builder()
-                .push(
-                    AnyBalances::new_builder()
-                        .set(AnyBalancesUnion::CKByteDistribution(
-                            CKByteDistribution::new_builder()
-                                .set(balances_array)
-                                .build(),
-                        ))
-                        .build(),
-                )
-                .build(),
-        )
-        .locked(
-            LockedBalances::new_builder()
-                .push(SubAlloc::new_builder().build())
-                .build(),
-        )
-        .build();
-    let channel_state = ChannelState::new_builder()
-        .channel_id(Byte32::zero())
-        .balances(balances)
-        .is_final(Bool::from_bool(true))
-        .version(10u64.pack())
-        .build();
-    let state_eth = convert_ckb_state(&channel_state);
-    println!("State: {:?}", state_eth);
-    let state_abi_encoded = state_eth.abi_encode();
-    println!("Encoded: {:?}", state_abi_encoded);
-    let msg_hash = ethereum_message_hash(&state_abi_encoded);
-    println!("Hash: {:?}", msg_hash);
-    verify_signature(&msg_hash, &sig_bytes, pubkey.as_slice()).expect("valid signature");
-}
-
-fn u128_le_as_uint128(v: u128) -> Uint128 {
-    let le = v.to_le_bytes();
-    // Uint128 expects exactly 16 bytes
-    Uint128::new_unchecked(molecule::bytes::Bytes::from(le.to_vec()))
-}
-
-fn eth_address_from_hex(s: &str) -> EthAddress {
-    let s = s.trim_start_matches("0x");
-    let raw = hex::decode(s).expect("valid hex address");
-    assert_eq!(raw.len(), 20);
-    let arr20: [Byte; 20] = raw
-        .into_iter()
-        .map(|b| Byte::from(b))
-        .collect::<Vec<_>>()
-        .try_into()
-        .unwrap();
-    EthAddress::new_builder().set(arr20).build()
-}
-#[test]
-fn test_cross_signature() {
-    // This tests the interoperability between the on-chain signature verification
-    // and the key generation & signing in the perun-ckb-backend's wallet.
-
-    let channel_id_bytes: [u8; 32] = [
-        44, 200, 243, 93, 27, 28, 68, 193, 103, 206, 252, 229, 221, 0, 109, 14, 208, 193, 32, 231,
-        123, 54, 49, 227, 99, 145, 72, 1, 70, 113, 99, 168,
-    ];
-
-    // Version
-    let version: u64 = 1;
-
-    // CKBytes balances (Uint64 each)
-    let ck_a: u64 = 8_000_000_000;
-    let ck_b: u64 = 18_000_000_000;
-
-    // One ETH asset:
-    let eth_chain_id: u128 = 1337; // will be taken LE as Uint128
-    let eth_addr_hex = "4E1D65a1E558029058903528140438802a6B5dfC";
-    let eth_a: u128 = 1_000_000_000_000_000_000u128; // 1e18
-    let eth_b: u128 = 0;
-
-    // This signature was generated by the wallet in the perun-ckb-backend
-    let sig_string = "304402200b8327fb275d132b5f70b1a578ca92127901c6d780911a742809940c11482bab022073904cdc55036ec1a39ed2fdddc0646eef8e4fb721a0253df68e56e8e830d74b";
-    let sig = hex::decode(sig_string).expect("decoding signature");
-    let sig_bytes: Bytes = sig.into();
-
-    // This public key was generated by the wallet in the perun-ckb-backend
-    let pubkey_string = "025a00821d5bd9a264c54f411d0a1a490fd0283eee6f9650ca09dbc4d65d173150";
-    let pubkey = hex::decode(pubkey_string).expect("decoding pubkey");
-    let pubkey_bytes: [Byte; 33] = pubkey
-        .iter()
-        .map(|x| (*x).into())
-        .collect::<Vec<Byte>>()
-        .try_into()
-        .unwrap();
-    SEC1EncodedPubKey::new_builder().set(pubkey_bytes).build();
-
-    let ck_dist = AnyBalancesUnion::CKByteDistribution(
-        CKByteDistribution::new_builder()
-            .set([ck_a.pack(), ck_b.pack()])
-            .build(),
-    );
-
-    // ETH asset & distribution
-    let chain_id_u128 = u128_le_as_uint128(eth_chain_id);
-    let eth_addr = eth_address_from_hex(eth_addr_hex);
-    let eth_asset = ETHAsset::new_builder()
-        .chain_id(chain_id_u128)
-        .asset_address(eth_addr)
-        .build();
-
-    let eth_dist = ETHDistribution::new_builder()
-        .nth0(u128_le_as_uint128(eth_a))
-        .nth1(u128_le_as_uint128(eth_b))
-        .build();
-
-    let eth_balances = AnyBalancesUnion::ETHBalances(
-        ETHBalances::new_builder()
-            .asset(eth_asset)
-            .distribution(eth_dist)
-            .build(),
-    );
-
-    // Locked: push a default SubAlloc (keeps Molecule happy; convert() ignores it)
-    let locked = LockedBalances::new_builder()
-        .push(SubAlloc::new_builder().build())
-        .build();
-
-    let balances = Balances::new_builder()
-        .assets(
-            Allocation::new_builder()
-                .push(AnyBalances::new_builder().set(ck_dist).build())
-                .push(AnyBalances::new_builder().set(eth_balances).build())
-                .build(),
-        )
-        .locked(locked)
-        .build();
-
-    let channel_state = ChannelState::new_builder()
-        .channel_id(Byte32::from_slice(&channel_id_bytes).unwrap())
-        .balances(balances)
-        .is_final(Bool::from_bool(true))
-        .version(version.pack())
-        .build();
-    let state_eth = convert_ckb_state(&channel_state);
-    println!("State: {:?}", state_eth);
-    let state_abi_encoded = state_eth.abi_encode();
-    println!("Encoded: {:?}", state_abi_encoded);
-    let msg_hash = ethereum_message_hash(&state_abi_encoded);
-    println!("Hash: {:?}", msg_hash);
-    verify_signature(&msg_hash, &sig_bytes, pubkey.as_slice()).expect("valid signature");
-}
-
-mod ckb_keys {
-    use crate::perun::Account;
-    use crate::perun::TestAccount;
-    use k256::ecdsa::SigningKey;
-    use k256::elliptic_curve::sec1::ToEncodedPoint;
-    use k256::PublicKey;
-    pub fn generate_ckb_keypair() -> (SigningKey, PublicKey) {
-        let acc_name = "alice".to_string();
-        let account = TestAccount::new_with_random_key(acc_name);
-        let pubkey = account.public_key();
-        let signing_key = account.sk.clone();
-        (signing_key, pubkey)
-    }
-
-    pub fn print_ckb_keypair() {
-        let (signing_key, public_key) = generate_ckb_keypair();
-        println!(
-            "CKB Private Key (hex): 0x{}",
-            hex::encode(signing_key.to_bytes())
-        );
-        println!(
-            "CKB Public Key  (hex): 0x{}",
-            hex::encode(public_key.to_encoded_point(false).as_bytes())
-        );
-    }
-}
-
-#[test]
-fn test_generate_ckb_keys() {
-    ckb_keys::print_ckb_keypair();
-}
 
 #[test]
 // TODO: Add mutator to channel state that can be passed to dispute, and close.
@@ -2937,4 +2727,3 @@ fn test_vc_happy_multi_asset_with_merge(
         },
     )
 }
-
